refactor(ui): split update status builders
